move some stuff over to the utils.py
