+++ src/pyIEEM/models/utils.py
@@ -1,13 +1,17 @@
 import os
 import numpy as np
 import pandas as pd
-from pyIEEM.data.utils import to_pd_interval, aggregate_contact_matrix, convert_age_stratified_property
+from datetime import datetime, timedelta
 from pyIEEM.models.models import SIR
+from pyIEEM.data.utils import to_pd_interval, aggregate_contact_matrix, convert_age_stratified_property
 
 abs_dir = os.path.dirname(__file__)
 
+#################################
+## Initialisation of the model ##
+#################################
 
-def initialize_SIR(country, age_classes, spatial=True, contact_type='absolute_contacts'):
+def initialize_model(country, age_classes, spatial=True, contact_type='absolute_contacts'):
 
     # Get model parameters
     # ====================
@@ -245,9 +249,7 @@ def get_epi_params(country, age_classes, spatial, contact_type):
     # ==================
 
     # infectivity
-    R0_desired = 3.3
-    
-    parameters = {'beta': 0.014}
+    parameters = {'beta': 0.024}
 
     # durations
     parameters.update({'alpha': 4.5,
@@ -302,3 +304,206 @@ def get_epi_params(country, age_classes, spatial, contact_type):
 
     return initial_states, parameters, coordinates
 
+########################################
+## Time-dependent Parameter Functions ##
+########################################
+
+def ramp_fun(t, t_start, l, N_old, N_new):
+    """
+
+    input
+    =====
+
+    t : timestamp
+        current date
+
+    t_start : timestamp
+        start of ramp
+
+    l : timestamp
+        length of ramp
+
+    N_old : float/np.array
+        old policy
+    
+    N_new : float/np.array
+        new policy
+
+    output
+    ======
+
+    N_t : float/np.array
+        interpolation between N_old and N_new
+    """
+    if t_start < t < t_start+timedelta(days=l):
+        return N_old + ((N_new-N_old)/l)*((t-t_start)/timedelta(days=1))
+    else:
+        return N_new
+
+from dateutil.easter import easter
+def is_Belgian_school_holiday(d):
+    """
+    A function returning 'True' if a given date is a school holiday or primary and secundary schools in Belgium
+    Tertiary education, which starts Sept. 12 and has exam periods in Dec-Jan and May-June are not considered.
+    
+    Input
+    -----
+    
+    d: datetime.datetime
+        Current simulation date
+    
+    Returns
+    -------
+    
+    is_Belgian_school_holiday: bool
+        True: date `d` is a school holiday for primary and secundary schools
+    """
+    
+    # Pre-allocate a vector containing the year's holiday weeks
+    holiday_weeks = []
+    
+    # Herfstvakantie
+    holiday_weeks.append(44)
+    
+    # Extract date of easter
+    d_easter = easter(d.year)
+    # Convert from datetime.date to datetime.datetime
+    d_easter = datetime(d_easter.year, d_easter.month, d_easter.day)
+    # Get week of easter
+    w_easter = d_easter.isocalendar().week
+
+    # Default logic: Easter holiday starts first monday of April
+    # Unless: Easter falls after 04-15: Easter holiday ends with Easter
+    # Unless: Easter falls in March: Easter holiday starts with Easter
+    if d_easter >= datetime(year=d.year,month=4,day=15):
+        w_easter_holiday = w_easter - 1
+    elif d_easter.month == 3:
+        w_easter_holiday = w_easter + 1
+    else:
+        w_easter_holiday = datetime(d.year, 4, (8 - datetime(d.year, 4, 1).weekday()) % 7).isocalendar().week
+    holiday_weeks.append(w_easter_holiday)
+    holiday_weeks.append(w_easter_holiday+1)
+
+    # Krokusvakantie
+    holiday_weeks.append(w_easter-6)
+
+    # Extract week of Christmas
+    # If Christmas falls on Saturday or Sunday, Christams holiday starts week after
+    w_christmas_current = datetime(year=d.year,month=12,day=25).isocalendar().week
+    if datetime(year=d.year,month=12,day=25).isoweekday() in [6,7]:
+        w_christmas_current += 1
+    w_christmas_previous = datetime(year=d.year-1,month=12,day=25).isocalendar().week
+    if datetime(year=d.year-1,month=12,day=25).isoweekday() in [6,7]:
+        w_christmas_previous += 1
+    # Christmas logic
+    if w_christmas_previous == 52:
+        if datetime(year=d.year-1, month=12, day=31).isocalendar().week != 53:
+            holiday_weeks.append(1)   
+    if w_christmas_current == 51:
+        holiday_weeks.append(w_christmas_current)
+        holiday_weeks.append(w_christmas_current+1)
+    if w_christmas_current == 52:
+        holiday_weeks.append(w_christmas_current)
+        if datetime(year=d.year, month=12, day=31).isocalendar().week == 53:
+            holiday_weeks.append(w_christmas_current+1)
+
+    # Define Belgian Public holidays
+    public_holidays = [
+        datetime(year=d.year, month=1, day=1),       # New Year
+        d_easter + timedelta(days=1),                # Easter monday
+        datetime(year=d.year, month=5, day=1),       # Labor day
+        d_easter + timedelta(days=40),               # Acension Day
+        datetime(year=d.year, month=7, day=21),      # National holiday
+        datetime(year=d.year, month=8, day=15),      # Assumption Mary
+        datetime(year=d.year, month=11, day=1),      # All Saints
+        datetime(year=d.year, month=11, day=11),     # Armistice
+        datetime(year=d.year, month=12, day=25),     # Christmas
+    ]
+    
+    # Logic
+    if ((d.isocalendar().week in holiday_weeks) | \
+            (d in public_holidays)) | \
+                ((datetime(year=d.year, month=7, day=1) <= d < datetime(year=d.year, month=9, day=1))):
+        return True
+    else:
+        return False
+
+def aggregate_simplify_contacts(contact_df, age_classes, demography, contact_type):
+    """
+    A function to demographically convert the contacts to the right age classes and simplify the (very large) pd.DataFrame `contact_df`
+
+    input
+    =====
+
+    contact_df: pd.DataFrame
+        contacts obtained from Beraud (2015)
+
+    age_classes: pd.IntervalIndex
+        age classes of the epidemiological model
+
+    demography: pd.Series
+        demography of country under study.
+
+    contact_type: str
+        'absolute_contacts'/'absolute_contacts_less_5_min' versus 'integrated_contacts'/'integrated_contacts_less_5_min'
+
+    output
+    ======
+
+    N_home: pd.Series
+        Home contacts. Index: [day_type, vacation, age_x, age_y]. Values: `contact_type`
+
+    N_leisure_private: pd.Series
+        Private leisure contacts. Idem.
+
+    N_leisure_public: pd.Series
+        Public leisure contacts. Idem.
+
+    N_school: pd.Series
+        School contacts. Idem.
+
+    N_work: pd.Series
+        Work contacts. Index: [sector, day_type, vacation, age_x, age_y]. Values: `contact_type`
+    """
+
+    # Determine if a demographic conversion is necessary (this is computationally demanding and takes approx. 1 min.)
+    if all(age_classes == pd.IntervalIndex.from_tuples([(0,5),(5,10),(10,15),(15,20),(20,25),(25,30),(30,35),(35,40),(40,45),(45,50),(50,55),(55,60),(60,65),(65,70),(70,75),(75,80),(80,120)], closed='left')):
+        convert = False
+    else:
+        convert = True
+
+    matrices=[]
+    for location in contact_df.index.get_level_values('location').unique().values:
+        if location not in ['SPC', 'work']:
+            # output dataframe
+            type_days = contact_df.index.get_level_values('type_day').unique().values
+            vacations = contact_df.index.get_level_values('vacation').unique().values
+            age_x = age_y = age_classes
+            iterables = [type_days, vacations, age_x, age_y]
+            contact = pd.Series(index=pd.MultiIndex.from_product(iterables, names=['type_day', 'vacation', 'age_x', 'age_y']), name=location, dtype=float)                           
+            for type_day in type_days:
+                for vacation in vacations:
+                    m = contact_df.loc[location, 'A', type_day, vacation, slice(None), slice(None)][contact_type]
+                    if convert:
+                        m = aggregate_contact_matrix(m, age_classes, demography)
+                    contact.loc[type_day, vacation, slice(None), slice(None)] = m.values
+            matrices.append(contact)
+
+        elif location == 'work':
+            # output dataframe
+            sectors = contact_df.index.get_level_values('sector').unique().values
+            type_days = contact_df.index.get_level_values('type_day').unique().values
+            vacations = contact_df.index.get_level_values('vacation').unique().values
+            age_x = age_y = age_classes
+            iterables = [sectors, type_days, vacations, age_x, age_y]
+            contact = pd.Series(index=pd.MultiIndex.from_product(iterables, names=['sector', 'type_day', 'vacation', 'age_x', 'age_y']), name=location, dtype=float)                           
+            for sector in sectors:
+                for type_day in type_days:
+                    for vacation in vacations:
+                        m = contact_df.loc[location, sector, type_day, vacation, slice(None), slice(None)][contact_type]
+                        if convert:
+                            m = aggregate_contact_matrix(m, age_classes, demography)
+                        contact.loc[sector, type_day, vacation, slice(None), slice(None)] = m.values
+            matrices.append(contact)
+
+    return (*matrices,)
